add checjAccessRight middlleware

+++ routes/api/feedback.js
@@ -1,7 +1,11 @@
 const express = require("express");
 const router = express.Router();
 const { controllerExceptionWrapper } = require("../../helpers/utils");
-const { validateBody, authUser } = require("../../middlewares");
+const {
+  validateBody,
+  authUser,
+  checkAccessRight,
+} = require("../../middlewares");
 const { sendFeedBackSchema } = require("../../helpers/schemas");
 const feedbackController = require("../../controllers/feedback");
 
@@ -14,16 +18,19 @@ router
   .get(
     "/all",
     authUser,
+    checkAccessRight(),
     controllerExceptionWrapper(feedbackController.getListOfFeedBack)
   )
   .get(
     "/:id",
     authUser,
+    checkAccessRight(),
     controllerExceptionWrapper(feedbackController.getCertainFeedback)
   )
   .delete(
     "/:id",
     authUser,
+    checkAccessRight(),
     controllerExceptionWrapper(feedbackController.deleteById)
   );
